backend/requirements.txt
Django==5.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
google-generativeai>=0.3.0
PyJWT>=2.10.1
